Add feature to allow Portal and BpkPopoverPortal to recieve custom style and className props
It's currently not easy to control the z-index of the portal when creating a new
bpk-component-popover component. Custom props are added to the BpkPopover component
and not the parent Portal component.

This change allows you to specify the portalClassName and portalStyle properties
which are added to the Portal component. This can be helpful when you need to
change the default style of the Portal (for example, when needing to manage the z-index).