Remove onClick wrapper for BpkButton
Tested this funcitonality in Safari, Chrome and FF and the bug doesn't
appear.

This issue was raised in 2016 and the issue closed in 2018 due to
inactivity.

It's likely this bug was fixed a long time ago so any risk from old
browsers still with this bug (if it was a browser bug) is very minimal