BpkButton: Improve onClick type to Mouse Event
SyntheticEvent is react's most generic kind of event, but specifying a
MouseEvent we get more properties. It also aligns with intellisense
of vscode

For keyboard users, onClick is still fired, but just some properties
of the event aren't populated, but they do exist.

If we wanted to responed to keyboard events, we would use the onKeyDown
prop