Fix bpkButton ts imports
Importing ts types should be done with the type keyword. When the
code is transpiled, these imports are stripped out.

If we don't add the keyword, the import of Props sticks around, and if
consumers evaluate the code at all, it will fail as the module has no
export named Props as it's already been stripped out.